changes form component to work better with antd
