+++ src/App.tsx
@@ -1,13 +1,16 @@
 import React from "react";
 import "./App.scss";
 
-import Carousel from "./Carousel";
-import FormComponent from "./FormComponent";
+import Carousel from "./components/Carousel";
+import FormComponent from "./components/FormComponent";
+
 
 const App: React.FC = () => {
   return (
     <div className="App">
-      <FormComponent />
+     
+     
+      <FormComponent/>
       <Carousel />
     </div>
   );
